Try background_color and theme_color in manifest
Chrome Beta 46 enables these two members in Web Application Manifest:

https://www.chromestatus.com/feature/5651543017652224
The background_color attribute in the JSON Web Application Manifest
allows web developers to provide a hint to Chrome indicating the
background color of their web application. This color is used as the
background of the application while its resources are fetched.

https://www.chromestatus.com/feature/5709006190411776
The theme_color attribute in the JSON Web Application Manifest allows
web developers to provide a global theme color which can be used
throughout their entire website. This color is used for the header in
the recents menu of Android.

Note: It should be noted that if a theme color is specified as a
meta-tag for a certain page (see
https://developers.google.com/web/updates/2014/11/Support-for-theme-color-in-Chrome-39-for-Android?hl=en),
then the information from the meta tag will override the global theme
color option in the manifest for that page.

http://www.color-hex.com/color/2196f3
Hex color #2196f3,  RGB value is (33,150,243).

https://airhorner.com/manifest.json